Provide more useful standard functionality
It doesn't feel right to dictate a standard policy of 'per minute' and
'per hour' periodic logging, so the MetricsConfigurationSection probably
shouldn't be part of this project. But removing it would require a major
version bump, so simply make a copy of the useful
PeriodicLogConfigurationElement at the top level.

The System.IO.Compression namespace offers support for ZIP archives, so
provide a sane default archiver implementation.

Allow for 'dynamic' environment value strings. These are usually
expected to be static but there is some value in permitting them to be
generated on-the-fly, since not all 'metrics' are numeric. Until the
Metrics JSON format allows for general-purpose string fields in the
hierarchy, we can output namespaced environment entries instead.

Bluewire.MetricsAdapter: 1.0.0 -> 1.1.0