Try to create (and recreate) counters for the correct instance
Performance counter instance names have no relation to PID. Determining
the correct instance name for a given process is fraught with peril and
race conditions, and instance names can change while the process runs.

This implements an alternative Metrics counter wrapper which tries to be
smart about instance names, looking for the one which corresponds to the
current PID and attempting to handle race conditions.

Assumes that the instance name for a given process is the same within a
group of counter categories, and that this applies to all counter
categories which we care about.